+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
  * cosmetics (future parser)
  * fix `STRIC_VARIABLES`
+ * update puppet files
 
 ## 22-08-2016 release 0.7.1
   * [BUGFIX] define multiple ssh keys